refactor: refactor key management and move address parsers to keys package
