+++ crypto/keys/methods.go
