+++ crypto/mpc/encoder.go
@@ -9,7 +9,29 @@ import (
 	"github.com/onsonr/sonr/crypto/keys"
 )
 
-type Role int
+// ╭───────────────────────────────────────────────────────────╮
+// │                    Exported Generics                      │
+// ╰───────────────────────────────────────────────────────────╯
+
+type (
+	Role        int                            // Role is the type for the role
+	Message     *protocol.Message              // Message is the protocol.Message that is used for MPC
+	Signature   *curves.EcdsaSignature         // Signature is the type for the signature
+	RefreshFunc interface{ protocol.Iterator } // RefreshFunc is the type for the refresh function
+	SignFunc    interface{ protocol.Iterator } // SignFunc is the type for the sign function
+)
+
+func GetIssuerDID(pk keys.PubKey) (string, string, error) {
+	addr, err := bech32.ConvertAndEncode("idx", pk.Bytes())
+	if err != nil {
+		return "", "", err
+	}
+	return fmt.Sprintf("did:sonr:%s", addr), addr, nil
+}
+
+// ╭───────────────────────────────────────────────────────────╮
+// │                  MPC Share Roles (Alice/Bob)              │
+// ╰───────────────────────────────────────────────────────────╯
 
 const (
 	RoleUnknown Role = iota
@@ -20,22 +42,8 @@ const (
 func (r Role) IsUser() bool      { return r == RoleUser }
 func (r Role) IsValidator() bool { return r == RoleValidator }
 
-// Message is the protocol.Message that is used for MPC
-type (
-	Message   *protocol.Message
-	Signature *curves.EcdsaSignature
-)
+// ╭───────────────────────────────────────────────────────────╮
+// │                      Keyshare Encoding                    │
+// ╰───────────────────────────────────────────────────────────╯
 
-// RefreshFunc is the type for the refresh function
-type RefreshFunc interface{ protocol.Iterator }
-
-// SignFunc is the type for the sign function
-type SignFunc interface{ protocol.Iterator }
-
-func GetIssuerDID(pk keys.PubKey) (string, string, error) {
-	addr, err := bech32.ConvertAndEncode("idx", pk.Bytes())
-	if err != nil {
-		return "", "", err
-	}
-	return fmt.Sprintf("did:sonr:%s", addr), addr, nil
-}
+type KeyShare string
